--- a/ocrd_browser/application.py
+++ b/ocrd_browser/application.py
@@ -1,8 +1,8 @@
 from gi.repository import Gio, Gtk, GLib, Gdk
 
-import pkg_resources
 from typing import List
 
+from importlib.metadata import entry_points
 from ocrd_browser.util.gtk import ActionRegistry
 from ocrd_browser.ui import MainWindow, AboutDialog, OpenDialog
 from ocrd_browser.view import ViewRegistry
@@ -36,7 +36,7 @@ class OcrdBrowserApplication(Gtk.Application):
         self.set_accels_for_action('view.zoom_to::width', ['<Ctrl>numbersign'])
         self.set_accels_for_action('view.zoom_to::page', ['<Ctrl><Alt>numbersign'])
 
-        for entry_point in pkg_resources.iter_entry_points('ocrd_browser_ext'):
+        for entry_point in entry_points().get('ocrd_browser_ext', []):
             (entry_point.load())(self)
 
         self.load_css()