--- a/ocrd_browser/application.py
+++ b/ocrd_browser/application.py
@@ -65,7 +65,7 @@ class OcrdBrowserApplication(Gtk.Application):
         open_windows: int = 0
         window: MainWindow
         for window in self.get_windows():
-            if isinstance(window, MainWindow) and window.close_confirm():  # type: ignore[unreachable]
+            if isinstance(window, MainWindow) and window.close_confirm():
                 window.destroy()
             else:
                 open_windows += 1