--- a/ocrd_browser/util/config.py
+++ b/ocrd_browser/util/config.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import os
 from collections import OrderedDict
 from configparser import ConfigParser
@@ -13,7 +14,7 @@ ConfigDict = MutableMapping[str, str]
 
 class _SubSettings:
     @classmethod
-    def from_section(cls, _name: str, section: ConfigDict) -> '_SubSettings':
+    def from_section(cls, _name: str, section: ConfigDict) -> _SubSettings:
         raise NotImplementedError('please override from_section')
 
     def validate(self) -> None:
@@ -28,7 +29,7 @@ class _FileGroups(_SubSettings):
         self.preferred_images = preferred_images
 
     @classmethod
-    def from_section(cls, _name: str, section: ConfigDict) -> '_FileGroups':
+    def from_section(cls, _name: str, section: ConfigDict) -> _FileGroups:
         preferred_images = section.get('preferredImages', 'OCR-D-IMG, OCR-D-IMG.*')
         return cls(
             [grp.strip() for grp in preferred_images.split(',')]
@@ -50,7 +51,7 @@ class _Tool(_SubSettings):
             raise ValueError('Could not locate executable "{}"'.format(executable))
 
     @classmethod
-    def from_section(cls, name: str, section: ConfigDict) -> '_Tool':
+    def from_section(cls, name: str, section: ConfigDict) -> _Tool:
         return cls(
             name[len(cls.PREFIX):],
             section['commandline'],
@@ -78,13 +79,13 @@ class Settings:
         return '{}({})'.format(self.__class__.__name__, repr(vars(self)))
 
     @classmethod
-    def get(cls) -> 'Settings':
+    def get(cls) -> Settings:
         if cls._settings is None:
             cls._settings = Settings.build_default()
         return cls._settings
 
     @classmethod
-    def build_default(cls, config_dirs: Optional[List[str]] = None, validate: bool = True) -> 'Settings':
+    def build_default(cls, config_dirs: Optional[List[str]] = None, validate: bool = True) -> Settings:
         if config_dirs is None:
             config_dirs = GLib.get_system_config_dirs() + [GLib.get_user_config_dir()]
             try:
@@ -97,7 +98,7 @@ class Settings:
         return cls.build_from_files(config_files, validate)
 
     @classmethod
-    def build_from_files(cls, files: List[str], validate: bool = True) -> 'Settings':
+    def build_from_files(cls, files: List[str], validate: bool = True) -> Settings:
         log = getLogger('ocrd_browser.util.config._Settings.build_from_files')
         config = ConfigParser()
         setattr(config, 'optionxform', lambda option: option)