--- a/ocrd_browser/view/__init__.py
+++ b/ocrd_browser/view/__init__.py
@@ -1,5 +1,4 @@
 from .base import View
-from .registry import ViewRegistry
 from .html import ViewHtml
 from .images import ViewImages
 from .text import ViewText
@@ -7,6 +6,7 @@ from .xml import ViewXml
 from .empty import ViewEmpty
 from .diff import ViewDiff
 from .page import ViewPage
+from .registry import ViewRegistry
 
 
 __all__ = ['View', 'ViewRegistry', 'ViewImages', 'ViewText', 'ViewXml', 'ViewHtml', 'ViewEmpty', 'ViewDiff', 'ViewPage']