--- a/ocrd_browser/view/registry.py
+++ b/ocrd_browser/view/registry.py
@@ -1,5 +1,6 @@
-from pkg_resources import EntryPoint, iter_entry_points
-from typing import Dict, Tuple, Optional
+from __future__ import annotations
+from importlib.metadata import entry_points
+from typing import Dict, Tuple, Optional, Type
 from .base import View
 
 ViewInfo = Tuple[type, str, str]
@@ -10,10 +11,9 @@ class ViewRegistry:
         self.views = views
 
     @classmethod
-    def create_from_entry_points(cls) -> 'ViewRegistry':
+    def create_from_entry_points(cls) -> ViewRegistry:
         views = {}
-        entry_point: EntryPoint
-        for entry_point in iter_entry_points('ocrd_browser_view'):
+        for entry_point in entry_points().get('ocrd_browser_view', []):
             view_class = entry_point.load()
             assert issubclass(view_class, View)
             label = view_class.label if hasattr(view_class, 'label') else view_class.__name__
@@ -24,5 +24,5 @@ class ViewRegistry:
     def get_view_options(self) -> Dict[str, str]:
         return {id_: label for id_, (view_class, label, description) in self.views.items()}
 
-    def get_view(self, id_: str) -> Optional[type]:
+    def get_view(self, id_: str) -> Optional[Type[View]]:
         return self.views[id_][0] if id_ in self.views else None