Gettting rid of pkg_resources brings a good performance improvement, from 1.178s to 0.864s in startup time on my machine
Contains also some less `import`s when not TYPE_CHECKING 

```
$ STARTUP_PROFILE=1 browse-ocrd 
         1389599 function calls (1321032 primitive calls) in 1.178 seconds

   Ordered by: internal time
   List reduced from 4716 to 20 due to restriction <20>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
39203/262    0.063    0.000    0.243    0.001 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/pkg_resources/_vendor/pyparsing.py:1370(_parseNoCache)
        1    0.063    0.063    0.130    0.130 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/application.py:53(do_activate)
      575    0.051    0.000    0.051    0.000 {built-in method marshal.loads}
       34    0.041    0.001    0.041    0.001 {built-in method _imp.create_dynamic}
        1    0.035    0.035    0.036    0.036 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/application.py:26(do_startup)
    20365    0.029    0.000    0.033    0.000 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/pkg_resources/_vendor/pyparsing.py:363(__new__)
      455    0.024    0.000    0.024    0.000 {built-in method builtins.compile}
        1    0.024    0.024    0.052    0.052 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/view/page.py:310(build)
1767/1717    0.022    0.000    0.105    0.000 {built-in method builtins.__build_class__}
     5075    0.021    0.000    0.048    0.000 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/pkg_resources/_vendor/packaging/version.py:261(__init__)
20365/19896    0.020    0.000    0.030    0.000 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/pkg_resources/_vendor/pyparsing.py:372(__init__)
        1    0.018    0.018    0.018    0.018 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/view/base.py:141(__init__)
 1125/220    0.017    0.000    0.046    0.000 /usr/lib/python3.8/sre_parse.py:493(_parse)
     6433    0.017    0.000    0.017    0.000 {built-in method posix.stat}
 4796/652    0.016    0.000    0.211    0.000 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/pkg_resources/_vendor/pyparsing.py:3547(parseImpl)
   156784    0.016    0.000    0.016    0.000 {built-in method builtins.isinstance}
       58    0.015    0.000    0.015    0.000 {method 'get_g_type' of 'gi.RegisteredTypeInfo' objects}
 4864/262    0.013    0.000    0.242    0.001 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/pkg_resources/_vendor/pyparsing.py:3397(parseImpl)
        1    0.013    0.013    0.013    0.013 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/util/gtk.py:65(WhenIdle)
     1320    0.011    0.000    0.043    0.000 <frozen importlib._bootstrap_external>:1498(find_spec)




$ STARTUP_PROFILE=1 browse-ocrd 
         791765 function calls (774883 primitive calls) in 0.864 seconds

   Ordered by: internal time
   List reduced from 4653 to 20 due to restriction <20>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      577    0.066    0.000    0.066    0.000 {built-in method marshal.loads}
        1    0.063    0.063    0.127    0.127 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/application.py:53(do_activate)
       34    0.047    0.001    0.047    0.001 {built-in method _imp.create_dynamic}
        1    0.037    0.037    0.047    0.047 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/application.py:26(do_startup)
      455    0.029    0.000    0.029    0.000 {built-in method builtins.compile}
1778/1727    0.024    0.000    0.095    0.000 {built-in method builtins.__build_class__}
        1    0.021    0.021    0.049    0.049 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/view/page.py:310(build)
        1    0.019    0.019    0.019    0.019 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/ocrd_browser/view/base.py:141(__init__)
 1127/219    0.017    0.000    0.046    0.000 /usr/lib/python3.8/sre_parse.py:493(_parse)
     6272    0.016    0.000    0.016    0.000 {built-in method posix.stat}
       57    0.015    0.000    0.015    0.000 {method 'get_g_type' of 'gi.RegisteredTypeInfo' objects}
      577    0.011    0.000    0.011    0.000 {built-in method io.open_code}
     1322    0.011    0.000    0.043    0.000 <frozen importlib._bootstrap_external>:1498(find_spec)
      916    0.010    0.000    0.013    0.000 /usr/lib/python3.8/sre_compile.py:276(_optimize_charset)
     2909    0.010    0.000    0.022    0.000 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/pkg_resources/_vendor/packaging/version.py:261(__init__)
        1    0.009    0.009    0.184    0.184 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/gi/overrides/Gio.py:39(run)
 1953/207    0.008    0.000    0.023    0.000 /usr/lib/python3.8/sre_compile.py:71(_compile)
    69525    0.007    0.000    0.008    0.000 {built-in method builtins.isinstance}
        2    0.007    0.003    0.007    0.003 /home/jk/Projekte/ocrd_all_2020/venv/lib/python3.8/site-packages/gi/_gtktemplate.py:175(init_template)
      457    0.006    0.000    0.006    0.000 {built-in method posix.listdir}
```