@kba 2022-06-15
That is indeed a signifcant improvement, thanks for sharing (and implementing). We'll investigate to do the same in core and dependent packages, cf https://github.com/OCR-D/core/issues/881